\name{AdvancedDataAnalysis-package}
\alias{AdvancedDataAnalysis-package}
\alias{AdvancedDataAnalysis}
\docType{package}
\title{
\packageTitle{AdvancedDataAnalysis}
}
\description{
\packageDescription{AdvancedDataAnalysis}
}

\author{
\packageAuthor{AdvancedDataAnalysis}

Maintainer: \packageMaintainer{AdvancedDataAnalysis}
}
\keyword{ package }
