% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Athletes.R
\docType{data}
\name{Athletes}
\alias{Athletes}
\title{Results of Athletes}
\format{A data frame with 12 rows and 3 columns.
\describe{
 \item{AthleteID}{Athlete identification.}
 \item{Time1}{Test result at time 1.}
 \item{Time2}{Test result at time 2.}
}}
\usage{
Athletes
}
\description{
Results' track of a group of athletes at two different times.
}
\keyword{datasets}
