% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table_with_percent.R
\name{freq_table_with_percent}
\alias{freq_table_with_percent}
\title{freq_table_with_percent}
\usage{
freq_table_with_percent(df, cat_variable)
}
\arguments{
\item{df}{A data frame}

\item{cat_variable}{The quoted column name that contains the categorical variable}
}
\description{
Generates a frequency table with the percentage of each group.
}
