% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lec4Ex1.R
\docType{data}
\name{Lec4Ex1}
\alias{Lec4Ex1}
\title{Lec4Ex1}
\format{A data frame
\describe{
\item{Group}{Randomised Control or Treatment}
\item{Result}{Blood Pressure}}}
\usage{
Lec4Ex1
}
\description{
Data set used in the example 1 of lecture 4 of the module MS4215 in the
2017 autumn semester at the University of Limerick.
}
\keyword{datasets}
