% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lec4Ex4.R
\docType{data}
\name{Lec4Ex4}
\alias{Lec4Ex4}
\title{Lec4Ex4}
\format{A data frame with 80 rows and 3 columns
\describe{
\item{ID}{Tennis player Identification}
\item{Injury}{If injured or not}
\item{WarmUp}{If warmed up or not}}}
\usage{
Lec4Ex4
}
\description{
Data of tennis players and whether or not they warmed up AND or NOT injured
}
\keyword{datasets}
