% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_SPSS_resource_2.R
\docType{data}
\name{Results_SPSS_resource_2}
\alias{Results_SPSS_resource_2}
\title{Quality of life data set}
\format{A data frame with 60 rows and 2 column
\describe{
\item{Group}{Either Randomised Control or Treament}
\item{QoL}{The quality of live measurement}}}
\usage{
Results_SPSS_resource_2
}
\description{
A data set containing a quality of live measurement of two groups of people,
one of them having received a treatment and the other a placebo.
}
\keyword{datasets}
