% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_independet_samples.R
\name{t_test_independent_samples}
\alias{t_test_independent_samples}
\title{t_test_independent}
\usage{
t_test_independent_samples(df, categorical, quantitative, confi = 0.95)
}
\arguments{
\item{df}{A data.frame}

\item{categorical}{The quoted column name of the column which contains the groups}

\item{quantitative}{The quoted column name of the column which contains the variable to be tested}

\item{confi}{The confidence level choosen}
}
\description{
Tests if the diference in the mean of a variable between two groups is statistically significant.
}
