% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chi_square.R
\name{test_chi_square}
\alias{test_chi_square}
\title{test_chi_square}
\usage{
test_chi_square(df, cat_1, cat_2)
}
\arguments{
\item{df}{A data frame.}

\item{cat_1}{The quoted column name which contains one of categorical variable.}

\item{cat_2}{The quoted column name which contains the other categorical
variable to be adressed.}
}
\value{
The same list as chisq.test would return plus the observed proportion.
}
\description{
It is a wrapper for the chisq.test. Two main diffenrece are that this function
evaluates if the expected values of each cells is greater than 5 before computing
test and it also adds a observed proportion table.
}
\examples{
test_chi_square(df = data, cat_1 = "Pclass", cat_2 = "Survived")

}
