% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_equality_paired.R
\name{test_equality_paired}
\alias{test_equality_paired}
\title{test_equality_paired}
\usage{
test_equality_paired(df, time1, time2, confi = 0.95,
  alternative = "two.sided")
}
\arguments{
\item{time1}{The quoted column name of one of the situations.}

\item{time2}{The quoted column name of the other situation.}

\item{confi}{The confidence level.}

\item{alternative}{The alternative hypothesis for the t test and wilcox test. The default
value is "two.sided", others options are "greater" and "less".}
}
\value{
A list containg the shapiro test of normality and the appropriate test of mean/median
equality.
}
\description{
Tests if the mean difference of the same group in different situation, e.g.
before and after a training, is statiscally different. In order to do so first
it is tested if the differen between the situations, e.g.again the time before
the training and the time after the training is normally distibuted. If it is
then a paired t test takes place, otherwise it is computed the Wilcoxon Signed-Rank
Test.
}
\details{
@param df A data frame containing the values of the different situations in different
 columns, example in one column the time before the training and in other column the time
 after the training.
}
\examples{

test_equality_paired(df = data, time1=score_before, time2=score_after,confi = 0.95, alternative = "two.sided")

}
