% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Titanic_data.R
\docType{data}
\name{Titanic_data}
\alias{Titanic_data}
\title{Titanic's passengers}
\format{A data frame with 891 rows and 11 columns
\describe{
\item{PassengerID}{Passanger Identification}
\item{Survived}{Survival}
\item{Pclass}{Passenger class}
\item{Gender}{Gender}
\item{Age}{Age (years)}
\item{SibSp}{Number of siblings on board}
\item{Parch}{Number of parent on board}
\item{Ticket}{Ticket Number}
\item{Fare}{Price of fare}
\item{Cabin}{Cabin number}
\item{Embarked}{Port of Embarkation}
}}
\usage{
Titanic_data
}
\description{
Contains data about the Titanic's passangers
}
\keyword{datasets}
