% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Travel_to_University.R
\docType{data}
\name{Travel_to_University}
\alias{Travel_to_University}
\title{Travel to University questionarY}
\format{A data frame with 50 lines and 5 columns
\describe{
  \item{id}{Student identification}
  \item{q1}{How did you get to the university?}
  \item{q2}{How long did the journey take?}
  \item{q3}{Did the journey take longer or less time than
  you expected?}
  \item{q4}{Why, if at all, did your journey take longer than
  expected?}
  \item{q5}{Is 9:30 course start time?}
}}
\usage{
Travel_to_University
}
\description{
A questionary about the students' travel to the University
}
\keyword{datasets}
