\name{AdvancedDataAnalysis-package}
\alias{AdvancedDataAnalysis-package}
\alias{AdvancedDataAnalysis}
\docType{package}
\title{
\packageTitle{AdvancedDataAnalysis}
}
\description{
\packageDescription{AdvancedDataAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AdvancedDataAnalysis}
\packageIndices{AdvancedDataAnalysis}
}
\author{
\packageAuthor{AdvancedDataAnalysis}

Maintainer: \packageMaintainer{AdvancedDataAnalysis}
}
\keyword{ package }
